Created a View folder and moved all the classes related there.
